fix(wide): round the generic products explicitly
With the generic lanes in registers, the arm64 compiler fused each
product with the following add or subtract: 388 FMADDS/FMSUBS in the
wide stages, a 1-ulp MulSub difference and a changed float witness on
the arm64 runner. An assignment does not stop fusion; an explicit
conversion does, so Mul now returns float32(a*b) per lane.

The FMA gate missed the fusions because the float32 forms carry an S
suffix that its word boundary rejected. The pattern accepts it now.

+++ wide_lane_generic.go
@@ -66,9 +66,15 @@ func (a laneW) Sub(b laneW) laneW {
 	return laneW{l0: a.l0 - b.l0, l1: a.l1 - b.l1, l2: a.l2 - b.l2, l3: a.l3 - b.l3}
 }
 
-// Mul returns the lane-wise product.
+// Mul returns the lane-wise product. The conversions round each product
+// explicitly, which forbids the compiler to fuse it with a later add.
 func (a laneW) Mul(b laneW) laneW {
-	return laneW{l0: a.l0 * b.l0, l1: a.l1 * b.l1, l2: a.l2 * b.l2, l3: a.l3 * b.l3}
+	return laneW{
+		l0: float32(a.l0 * b.l0),
+		l1: float32(a.l1 * b.l1),
+		l2: float32(a.l2 * b.l2),
+		l3: float32(a.l3 * b.l3),
+	}
 }
 
 // Div returns the lane-wise quotient, rounded once like the scalar division.
